Add unit test, dummy out new showdown format
no idea how soon psim will make the change on the public client, so just leave it for whenever it changes in the future.
I don't think it's worthwhile to expose it via temporary enum value.

Rename .Preview to .Brief since it's not a "feature-preview", it's a briefing/list of properties.
Add unit test to test translations and all lines recognized a<->b (yes it doesn't test EVERY token, but for a given set the test works)