DP - Add Spear Pillar related flags
Explanation of the various flags:
0350 - Set ON after beating/catching Dialga/Palkia in Spear Pillar during the story. When ON, the man blocking access to Route 222 due to a blackout at Sunyshore City is no longer there.
0351 - Set ON after catching Dialga/Palkia. Dialga/Palkia will not reappear after getting to the Hall of Fame.
0482 - Set ON after beating/catching Dialga/Palkia, set OFF after making it respawn by getting to the Hall of Fame without having caught it. The Dialga/Palkia NPC at the top of Spear Pillar will no longer be present.
1023 - Self-explanatory.
2433 - Set ON after beating/catching Dialga/Palkia during the story. Unlock the 2nd level of rematch against trainers. Applies to Platinum as well, so I added it.

I haven't looked into the Platinum flags since it was out of my scope, and that it would require to resume from scratch the search on this since the Spear Pillar story is completely different there, so I just checked quickly those that could be easy to check, which ended up being the rematch flag.

Misc flags that could be added if wanted:
0685 - Set ON after beating/catching Dialga/Palkia during the story. The Galaxy grunts will be gone from Mount Coronet.
0486 - Set ON after beating/catching Dialga/Palkia during the story. The Galaxy grunts will be gone from Spear Pillar.
0487 - Same as 0486, but for the Commanders.
0488 - Same as 0486, but for Cyrus.
0492 - Same as 0486, but for Uxie.
0493 - Same as 0486, but for Mesprit.
0494 - Same as 0486, but for Azelf.
I added none of them due to being heavily tied to the story (cutscene stuff), and making one of those disappear outside of the Mount Coronet grunts while not having finished the Spear Pillar script will lead to a crash (same if 0484 or 0485 is ON while doing the story at Spear Pillar).

Now, why did I searched those flags:
Diamond and Pearl have a glitch known as Tweaking Trick, allowing to do a lot of things, like getting to legendaries/mythical earlier without needing the key item, unlocking Fly locations earlier, and skipping some story events. Among those possibilities is getting to the Hall of Fame at 2 badges (https://www.youtube.com/watch?v=SlU_p1FEjb0). While it works and can be interesting to do for players, granting the player with a star, allowing to catch the lake guardians without having visited the Spear Pillar and done the story, allowing the National Dex to be unlocked earlier and Battle Frontier to be visited earlier, it comes with the flaw of breaking the story elements on the top of Spear Pillar. This leads to 1) some Galaxy members like Cyrus not being able to be challenged at Spear Pillar, being frozen in place and not doing anything, 2) Galaxy members in Mount Coronet and Spear Pillar not being able to leave since the Spear Pillar event cannot be completed, 3) prevent the blackout at Sunyshore City to finish, preventing the access to Route 222 without doing additional glitches or save editing, and 4) rematches being locked to those after beating Fantina since the level of rematch unlocked by beating the Spear Pillar content cannot be unlocked in this glitched state.
The documentation of those flags allow players having done this bug to avoid all those issues without having to start a new save file, either by setting the Dialga/Palkia variable to 0, setting the flag 482 to OFF (only ON if you caught Dialga/Palkia), and then completing the Spear Pillar normally, or by setting the flags 0350, 0482, 0486~0488, 0492~0494, 0685, and 2433 to ON while setting the Dialga/Palkia variable to 4.

Miscellaneous changes:
- Fix Accessible being written as Accessable
- Fix Jennifer being written as Jennnifer
- Fix the Korean translation not being up to date on the lines in the Platinum flags

Questions about this pull request, now:
- Is it all good to add those flags? (no idea what is a flag that should be added or not) Should I also add the flags mentioned in this comment, or are they not considered important?
- Is there a proper order, and proper letter to put for the flags? I had no real idea what was the formatting, so I tried to do what I could but it's a bit weird.