HGSS - Add flags tied to the Slowpoke Well
- Add more info about 0350 (also allows to get the Dialga/Palkia entry at Celestic Town when ON)
- Edit the type of flag used by those added in the previous commit per the documentation
- Add flags tied to the Slowpoke Well event

The Slowpoke Well flags are once more added for players having exploited Tweaking Trick during their playthrough, leading to issues later on and needing save editing (tweak in question: https://youtu.be/KK4GScMdcCc).
If a player exploited it during their playthrough, they will not be able to access Slowpoke Well without doing glitches/save editing, and Kurt's granddaughter will continue speaking as if the Slowpoke weren't rescued.
For those affected, setting the flags 0119 and 0415 ON should be all good.

Documentation about those flags:
They are all set ON when interacting with Kurt for the first time in his house.
0119 - When ON, Kurt's Granddaughter text will change, and she will worry about Kurt going to the Slowpoke Well instead of being worried about the Slowpoke. If this flag is OFF after finishing the Slowpoke Well quest, she will continue using her worried about Slowpoke line instead of her normal text.
0414 - When ON, Kurt will no longer be in his house. Set OFF when completing the Slowpoke Well quest.
0415 - When ON, the Rocket grunt blocking access to the Slowpoke Well will be gone.