Add zipped save file r/w
Not happy that zipping the file is the solution for some homebrew apps, but it is what it is.

No need to select which file; it's always one file in the zip, and never multiple.
When exporting, if it originated from a zip, grab the original then update it with the revised contents.

Closes #4564

Co-Authored-By: Chris Dailey <nitz@users.noreply.github.com>