Update WB8.cs
Manaphy eggs have MetLocation = Default8bNone (65535), which requires MetDate to be 0/0/0 according to the legality validator. Previously, the MetDate was being set to the distribution window date for all Pokémon, causing eggs to fail validation.

Changes:
- Clear MetDate components (Year/Month/Day) for Manaphy eggs after setting MetLocation to Default8bNone
- Skip setting MetDate for Manaphy eggs in the date assignment logic

This fixes the "Met Date is not a valid calendar date" error for Manaphy egg distributions (CardID 9026) in SysBot.NET projects.